requirements.txt
altair==4.2.2
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
archspec @ file:///croot/archspec_1697725767277/work
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.6.2
boltons @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/boltons_1699240838368/work
Brotli @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/brotli-split_1699237319520/work
cachetools==5.5.2
certifi==2025.6.15
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.2.2
conda @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_104cbbdti1/croot/conda_1708369123783/work
conda-content-trust @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/conda-content-trust_1699283351805/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1702997573971/work/src
conda-package-handling @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_3dblbw3die/croot/conda-package-handling_1700254361862/work
conda_package_streaming @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/conda-package-streaming_1699238858289/work
contourpy==1.2.1
cryptography==42.0.4
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
descope==1.6.5
distro @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/distro_1701803554172/work
dnspython==2.6.1
email-validator==2.0.0.post2
entrypoints==0.4
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.18.0
Flask==2.3.3
fonttools==4.53.1
fqdn==1.5.1
fsspec==2025.3.2
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==1.34.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.30.2
idna==3.7
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.4.0
json5==0.9.25
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
libmambapy @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/mamba-split_1701805498280/work/libmambapy
liccheck==0.9.2
lxml==5.4.0
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_26a7xsau78/croot/menuinst_1706732941263/work
mistune==3.0.2
narwhals==1.41.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.82.0
outcome==1.3.0.post0
overrides==7.7.0
packaging @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/packaging_1699237023876/work
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pdoc3==0.10.0
pexpect==4.9.0
pillow==10.3.0
platformdirs @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/platformdirs_1701803010714/work
playwright==1.52.0
pluggy @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pluggy_1699237243543/work
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pycosat_1699241479669/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.18.0
PyJWT==2.7.0
Pympler==1.1
PyMuPDF==1.26.1
pyparsing==3.1.2
PyPDF2==3.0.1
PySocks @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pysocks_1699237568675/work
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.4.16
requests==2.32.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.20.0
rsa==4.9.1
ruamel.yaml @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/ruamel.yaml_1699241608430/work
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
selenium==4.33.0
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==68.2.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.22.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.3.0
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
tornado==6.4.1
tqdm @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/tqdm_1701805884490/work
traitlets==5.14.3
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
truststore @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/truststore_1701805928787/work
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2024.1
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.4.0
validators==0.35.0
wcwidth==0.2.13
webcolors==24.6.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.41.2
widgetsnbextension==4.0.11
wsproto==1.2.0
youtube-transcript-api==1.0.3
zipp==3.23.0
zstandard @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/zstandard_1699238791645/work
